pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
debug = "backend.main:debug"
start = "backend.main:start"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
pre-commit = "^3.6.0"
mypy = "^1.8.0"
ruff = "^0.1.14"
black = "^23.12.1"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.28"}
httpx = "^0.26.0"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"
uvicorn = "^0.27.0"
sqlmodel = "^0.0.14"
bcrypt = "^4.1.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
sqlalchemy-stubs = "^0.4"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
